test: validate Gateway API v1.4 and v1.6
