+++ test/gateway/common/go.mod
@@ -1,0 +1,3 @@
+module github.com/alibaba/higress/v2/test/gateway/common
+
+go 1.24.0
