+++ tools/hack/run-gateway-api-conformance.sh
@@ -3,17 +3,20 @@
 set -euo pipefail
 
 GATEWAY_CLASS="${GATEWAY_CLASS:-higress}"
+GATEWAY_API_VERSION="${GATEWAY_API_VERSION:-v1.6.0}"
+CONFORMANCE_TEST_DIR="${GATEWAY_CONFORMANCE_TEST_DIR:-test/gateway/v1.6}"
 SUPPORTED_FEATURES="${GATEWAY_CONFORMANCE_SUPPORTED_FEATURES:-Gateway,HTTPRoute,TLSRoute,GRPCRoute,TCPRoute,ReferenceGrant}"
 CONFORMANCE_PROFILES="${GATEWAY_CONFORMANCE_PROFILE:-GATEWAY-HTTP,GATEWAY-TLS,GATEWAY-GRPC,GATEWAY-TCP}"
 REPORT="${GATEWAY_CONFORMANCE_REPORT:-out/gateway-api-conformance/report.yaml}"
 CONTACT="${GATEWAY_CONFORMANCE_CONTACT:-https://github.com/alibaba/higress/issues}"
 VERSION="${HIGRESS_CONFORMANCE_VERSION:-$(git rev-parse HEAD)}"
 ALLOW_CRDS_MISMATCH="${GATEWAY_CONFORMANCE_ALLOW_CRDS_MISMATCH:-false}"
+SUPPORTS_TEST_CLEANUP="${GATEWAY_CONFORMANCE_SUPPORTS_TEST_CLEANUP:-true}"
 CLEANUP_TEST_RESOURCES="${GATEWAY_CONFORMANCE_CLEANUP_TEST_RESOURCES:-true}"
 RUN_TEST="${GATEWAY_CONFORMANCE_RUN_TEST:-}"
 TEST_PARALLEL="${GATEWAY_CONFORMANCE_PARALLEL:-1}"
 KIND_CLUSTER_NAME="${KIND_CLUSTER_NAME:-higress}"
-CONFORMANCE_IMAGE="${HIGRESS_CONFORMANCE_TEST_IMAGE:-higress-gateway-conformance:v1.6-local}"
+CONFORMANCE_IMAGE="${HIGRESS_CONFORMANCE_TEST_IMAGE:-higress-gateway-conformance:${GATEWAY_API_VERSION}-local}"
 KUBECONFIG="${KUBECONFIG:-${HOME}/.kube/config}"
 
 if [[ ! -r "${KUBECONFIG}" ]]; then
@@ -42,7 +45,20 @@ esac
 WORK_DIR="$(mktemp -d)"
 trap 'rm -rf "${WORK_DIR}"' EXIT
 
-CGO_ENABLED=0 GOOS=linux GOARCH="${GOARCH}" go -C test/gateway test -c -o "${WORK_DIR}/gateway-conformance.test" .
+if [[ ! -f "${CONFORMANCE_TEST_DIR}/go.mod" ]]; then
+  echo "Gateway API conformance test module not found: ${CONFORMANCE_TEST_DIR}" >&2
+  exit 1
+fi
+if MODULE_VERSION="$(go -C "${CONFORMANCE_TEST_DIR}" list -m -f '{{.Version}}' sigs.k8s.io/gateway-api/conformance 2>/dev/null)"; then
+  :
+else
+  MODULE_VERSION="$(go -C "${CONFORMANCE_TEST_DIR}" list -m -f '{{.Version}}' sigs.k8s.io/gateway-api)"
+fi
+if [[ "${MODULE_VERSION}" != "${GATEWAY_API_VERSION}" ]]; then
+  echo "Gateway API test module version ${MODULE_VERSION} does not match requested version ${GATEWAY_API_VERSION}" >&2
+  exit 1
+fi
+CGO_ENABLED=0 GOOS=linux GOARCH="${GOARCH}" go -C "${CONFORMANCE_TEST_DIR}" test -c -o "${WORK_DIR}/gateway-conformance.test" .
 KIND_NODE_IMAGE="$(docker inspect "${KIND_NODE}" --format '{{.Config.Image}}')"
 docker build -q \
   --build-arg "KIND_NODE_IMAGE=${KIND_NODE_IMAGE}" \
@@ -70,10 +86,12 @@ ARGS=(
   "--contact=${CONTACT}"
   --mode=default
   --cleanup-base-resources=false
-  "--cleanup-test-resources=${CLEANUP_TEST_RESOURCES}"
   "--allow-crds-mismatch=${ALLOW_CRDS_MISMATCH}"
   "--report-output=/report/${REPORT_NAME}"
 )
+if [[ "${SUPPORTS_TEST_CLEANUP}" == "true" ]]; then
+  ARGS+=("--cleanup-test-resources=${CLEANUP_TEST_RESOURCES}")
+fi
 if [[ -n "${RUN_TEST}" ]]; then
   ARGS+=("--run-test=${RUN_TEST}")
 fi
